+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.3] - 2026-01-25
+
+### Added
+- Chocolatey dependency on Python 3.11+.
+
+### Changed
+- Chocolatey install script now ensures Python is available.
+
 ## [1.0.2] - 2026-01-25
 
 ### Added